tests/wpt/web-platform-tests/entries-api/interfaces-manual.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Entries API: IDL Tests</title>
<link rel=help href="https://wicg.github.io/entries-api/#idl-index">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/WebIDLParser.js"></script>
<script src="/resources/idlharness.js"></script>
<script src="support.js"></script>
<script>
'use strict';

entry_test((t, entry, item) => {
  assert_true(entry.isDirectory);
  Promise.all([
    getEntriesAsPromise(entry),
    fetch('/interfaces/entries-api.idl').then(r => r.text())
  ]).then(t.step_func(([entries, idls]) => {
    window.samples = {
      item: item,
      dirEntry: entries.filter(entry => entry.isDirectory)[0],
      fileEntry: entries.filter(entry => entry.isFile)[0],
      fileSystem: entry.filesystem,
    };

    const idl_array = new IdlArray();

    // https://w3c.github.io/FileAPI/#dfn-file
    idl_array.add_untested_idls('[Exposed=(Window,Worker)] interface File {};');

    // https://html.spec.whatwg.org/multipage/forms.html#htmlinputelement
    idl_array.add_untested_idls('interface HTMLInputElement {};');

    // https://html.spec.whatwg.org/multipage/interaction.html#datatransferitem
    idl_array.add_untested_idls('interface DataTransferItem {};');

    idl_array.add_idls(idls);

    idl_array.add_objects({
      File: ['new File([], "example.txt")'],
      HTMLInputElement: ['document.createElement("input")'],
      DataTransferItem: ['samples.item'],
      FileSystemEntry: [],
      FileSystemDirectoryEntry: ['samples.dirEntry'],
      FileSystemDirectoryReader: ['samples.dirEntry.createReader()'],
      FileSystemFileEntry: ['samples.fileEntry'],
      FileSystem: ['samples.fileSystem'],
    });

    idl_array.test();
    t.done();
  }));
}, 'Entries API: IDL');
</script>
